refactor(inv): remove redundant components
