+++ frontend/src/apps/inventory/components/InventoryList.tsx
