+++ frontend/src/apps/inventory/components/LowStockAlerts.tsx
