+++ frontend/src/apps/inventory/components/ProductFilters.tsx
