+++ frontend/src/apps/inventory/components/ProductRow.tsx
