+++ frontend/src/apps/inventory/components/ProductTable.tsx
