+++ frontend/src/apps/inventory/components/ProductTags.tsx
