+++ frontend/src/apps/inventory/components/index.tsx
@@ -1,0 +1,6 @@
+export { default as ProductFilters } from './ProductFilters';
+export { default as ProductTable } from './ProductTable';
+export { default as ProductRow } from './ProductRow';
+export { default as ProductTags } from './ProductTags';
+export { default as InventoryList } from './InventoryList';
+export { default as LowStockAlerts } from './LowStockAlerts';
