+++ frontend/src/apps/inventory/hooks/index.tsx
@@ -1,0 +1,1 @@
+export { useProducts } from './useProducts';
