+++ frontend/src/apps/inventory/hooks/useProducts.ts
@@ -1,8 +1,8 @@
 // This hook manages product data and filtering functionality
 
 import { useState, useEffect } from "react";
-import type { Product, ProductType, ProductBrand, ProductMaterial, Filters } from "@/types/product";
-import { fetchProducts, fetchTypes, fetchBrands, fetchMaterials } from "@/services/productService";
+import type { Product, ProductType, ProductBrand, ProductMaterial, Filters } from "../types/product";
+import { fetchProducts, fetchTypes, fetchBrands, fetchMaterials } from "../services/productService";
 
 export const useProducts = () => {
   const [products, setProducts] = useState<Product[]>([]);
