+++ frontend/src/apps/inventory/index.tsx
@@ -1,0 +1,20 @@
+export {
+  default as ProductOverview,
+  default as ArchivedProducts,
+  default as ProductForm
+} from './pages';
+
+export {
+  ProductFilters,
+  ProductTable,
+  ProductRow,
+  ProductTags,
+  InventoryList,
+  LowStockAlerts
+} from './components';
+
+export { useProducts } from './hooks';
+
+export { toggleProductArchive } from './services';
+
+export type * from './types/product';
