+++ frontend/src/apps/inventory/pages/index.tsx
@@ -1,0 +1,3 @@
+export { default as ProductOverview } from './ProductOverview';
+export { default as ArchivedProducts } from './ArchivedProducts';
+export { default as ProductForm } from './ProductForm';
