+++ frontend/src/apps/inventory/services/index.tsx
@@ -1,0 +1,1 @@
+export { toggleProductArchive } from './productService';
