+++ frontend/src/apps/inventory/services/productService.ts
