+++ frontend/src/apps/inventory/types/product.ts
